Strip dynamic routing param
